requirements.txt
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.7.1
seaborn==0.13.2
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.3
PyYAML==6.0.2
joblib==1.5.3
pytest==8.3.4
gunicorn==23.0.0
